Get information and help about R objects

Help obtained with this function is wider than with help(). If a man page is not found, it suggests related topics. If an object is an S3 generic function, it also lists all its known methods. Also, one can track the help page of an object even if its name is changed, by using the src or srcfile attribute of the object's comment. By the way, if the object has a comment, it is also displayed. This can be used as a quick and dirty way to provide short hints to custom objects. Finally, it is possible to track down the source of an object into a file with the srcfile attribute of its comment. In this case, it is the source file that is displayed. So, you can also further document your custom objects easily in their source files!

about(topic, ...)

`?`(type, topic)

Arguments

topic
The name of an object, or the topic to search for, if this is not the name of a known object.

...
Further arguments passed to help().

type
First argument to ?. If it is a dot, like .?topic, the second argument is a topic passed to the about() function. Otherwise, it is the first argument to restrict help pages, like class, methods, or method. See examples for how to use it.

Value

A string with the location of all objects named topic are found is returned invisibly.

See also

Examples

about("nonexisting") # Not found on search path, but help pages

#> 'nonexisting' not found

#> Searching keyword in all R help pages for ''...

about("htgdsfgfdsgf") # Not found anywhere

#> 'htgdsfgfdsgf' not found

#> Searching keyword in all R help pages for ''...

#> No vignettes or demos or help files found with alias or concept or
#> title matching ‘htgdsfgfdsgf’ using fuzzy matching.

#library(tidyverse)
#about("group_by") # Just one page
#about("filter") # Several items
about("stats::filter") # OK

#> Displaying help page...

#about("dplyr::filter") # OK too
about("base::filter") # Not found there

#> 'filter' not found in package 'base', do you mean?

#> stats::filter, base::Filter

#> Searching keyword in all R help pages for 'base'...

# Objects with comment: print comment
vec <- structure(1:10, comment = "A simple vector")
about("vec")

#> 'vec' not found, do you mean?

#> base::as.data.frame.vector, base::as.vector, base::as.vector.factor, base::is.vector, base::mat.or.vec, methods::removeClass, utils::RweaveChunkPrefix, base::vector, base::Vectorize

#> Searching keyword in all R help pages for ''...

# If there is a srcfile attribute in the comment, also display the file
# Hint: integrate some help in the header!
#library(data)
#(iris <- read(data_example("iris.csv")))
#about("iris")
# If the comment has a src attribute, change the topic to that one
#urchin <- read("urchin_bio", package = "data")
#about("urchin")
.?filter

#> Displaying help page...

.?stats::filter

#> Displaying help page...