Get a completion list for a R code fragment

Returns names of objects/arguments/namespaces matching a code fragment.

completion(
  code,
  pos = nchar(code),
  min.length = 2,
  print = FALSE,
  types = c("default", "scintilla"),
  addition = FALSE,
  sort = TRUE,
  what = c("arguments", "functions", "packages"),
  description = FALSE,
  max.fun = 100,
  skip.used.args = TRUE,
  sep = "\n",
  field.sep = "\t"
)

Arguments

code
A partial R code to be completed.

pos
The position of the cursor in this code.

min.length
The minimal length in characters of code required before the completion list is calculated.

print
Logical, print result and return invisibly. See details.

types
A named list giving names of types. Set to NA to give only names. See details.

addition
Should only addition string be returned?

sort
Do we sort the list of completions alphabetically?

what
What are we looking for? Allow to restrict search for faster calculation.

description
Do we describe items in the completion list (could be slow)?

max.fun
In the case where we describe items, the maximum number of functions to process (if longer, no description is returned for function) because it can be very slow otherwise.

skip.used.args
Logical, if completion is within function arguments, should the already used named arguments be omitted?

sep
The separator to use between returned items.

field.sep
Character string to separate fields for each entry.

Value

If types == NA and description = FALSE, a character vector giving the completions, otherwise a data frame with two columns: 'completion', and 'type' when description = FALSE, or with four columns: 'completion', 'type', 'desc' and 'context' when description = TRUE.
Attributes:
attr(, "token") - a completed token.
attr(, "triggerPos") - number of already typed characters.
attr(, "fguess") - name of guessed function.
`attr(, "isFirstArg")`` - is this a first argument?

Details

The completion list is context-dependent, and it is calculated as if the code was entered at the command line.

If the code ends with $ or [[, then the function look for items in a list or data.frame whose name is the last identifier.

If the code ends with @, then the function look for slots of the corresponding S4 object.

If the code ends with ::, then it looks for objects in a namespace.

If the code ends with a partial identifier name, the function returns all matching keywords visible from .GlobalEnv.

If the code is empty or parses into an empty last token, the list of objects currently in the global environment is returned.

Note

Take care: depending on the context, the completion list could be incorrect (but it should work for code entered at the command line). For instance, inside a function call, the context is very different, and arguments and local variables should be returned instead. This may be implemented in the future, but for now, we focus on completion that should be most useful for novice useRs that are using R expressions entered one after the other at the R console or in a script (and considering the script is run or sourced line after line in R).

There are other situations where the completion can be calculated, see the help of rc.settings().

If print == TRUE, results are returned invisibly, and printed in a form: triggerPos newline completions separated by sep.

If types are supplied, a completion will consist of name and type, separated by type.sep. types may me a vector of length 5, giving the type codes for "function", "variable", "environment", "argument" and "keyword". If types == "default", above type names are given; types == "scintilla" will give numeric codes that can be used with "scintilla.autoCShow" function (e.g., with the SciViews-K Komodo Edit plugin).

See also

Examples

# A data frame
data(iris)
completion("item <- iris$")

#>          completion     type
#> 3 iris$Petal.Length variable
#> 4  iris$Petal.Width variable
#> 1 iris$Sepal.Length variable
#> 2  iris$Sepal.Width variable
#> 5      iris$Species variable

completion("item <- iris[[")

#>             completion     type
#> 3 iris[["Petal.Length" variable
#> 4  iris[["Petal.Width" variable
#> 1 iris[["Sepal.Length" variable
#> 2  iris[["Sepal.Width" variable
#> 5      iris[["Species" variable

# An S4 object
setClass("track", representation(x = "numeric", y = "numeric"))
t1 <- new("track", x = 1:20, y = (1:20)^2)
completion("item2 <- t1@")

#>   completion     type
#> 1        t1@ variable

# A namespace
completion("utils::", description = TRUE)

#>                                       completion     type desc context
#> 65                                      utils::? variable        utils
#> 26                           utils::.DollarNames variable        utils
#> 131                               utils::.romans variable        utils
#> 55                      utils::.RtangleCodeLabel variable        utils
#> 129                            utils::.S3methods variable        utils
#> 190                                 utils::adist variable        utils
#> 28                                  utils::alarm variable        utils
#> 214                               utils::apropos variable        utils
#> 48                               utils::aregexec variable        utils
#> 12                           utils::argsAnywhere variable        utils
#> 219                             utils::as.person variable        utils
#> 110                         utils::as.personList variable        utils
#> 67                          utils::as.relistable variable        utils
#> 215                              utils::as.roman variable        utils
#> 108                           utils::asDateBuilt variable        utils
#> 145                              utils::askYesNo variable        utils
#> 205                                utils::aspell variable        utils
#> 212                utils::aspell_package_C_files variable        utils
#> 117                utils::aspell_package_R_files variable        utils
#> 1                 utils::aspell_package_Rd_files variable        utils
#> 21               utils::aspell_package_vignettes variable        utils
#> 183 utils::aspell_write_personal_dictionary_file variable        utils
#> 201                   utils::assignInMyNamespace variable        utils
#> 24                      utils::assignInNamespace variable        utils
#> 72                     utils::available.packages variable        utils
#> 156                              utils::bibentry variable        utils
#> 83                              utils::browseEnv variable        utils
#> 203                             utils::browseURL variable        utils
#> 100                       utils::browseVignettes variable        utils
#> 95                             utils::bug.report variable        utils
#> 94                         utils::capture.output variable        utils
#> 80                           utils::changedFiles variable        utils
#> 30                              utils::checkCRAN variable        utils
#> 199                      utils::chooseBioCmirror variable        utils
#> 138                      utils::chooseCRANmirror variable        utils
#> 175                              utils::citation variable        utils
#> 128                                  utils::cite variable        utils
#> 73                             utils::citeNatbib variable        utils
#> 140                              utils::citEntry variable        utils
#> 167                             utils::citFooter variable        utils
#> 52                              utils::citHeader variable        utils
#> 142                          utils::close.socket variable        utils
#> 146                                 utils::combn variable        utils
#> 109                        utils::compareVersion variable        utils
#> 101                           utils::contrib.url variable        utils
#> 36                           utils::count.fields variable        utils
#> 61                          utils::CRAN.packages variable        utils
#> 37                            utils::create.post variable        utils
#> 192                                  utils::data variable        utils
#> 171                            utils::data.entry variable        utils
#> 174                             utils::dataentry variable        utils
#> 119                                    utils::de variable        utils
#> 14                               utils::de.ncols variable        utils
#> 50                             utils::de.restore variable        utils
#> 47                               utils::de.setup variable        utils
#> 170                             utils::debugcall variable        utils
#> 137                              utils::debugger variable        utils
#> 115                                  utils::demo variable        utils
#> 213                         utils::download.file variable        utils
#> 136                     utils::download.packages variable        utils
#> 33                            utils::dump.frames variable        utils
#> 116                                  utils::edit variable        utils
#> 176                                 utils::emacs variable        utils
#> 191                               utils::example variable        utils
#> 162                             utils::file_test variable        utils
#> 51                              utils::file.edit variable        utils
#> 97                           utils::fileSnapshot variable        utils
#> 187                                  utils::find variable        utils
#> 158                           utils::findLineNum variable        utils
#> 53                                    utils::fix variable        utils
#> 189                        utils::fixInNamespace variable        utils
#> 106                         utils::flush.console variable        utils
#> 184                              utils::formatOL variable        utils
#> 8                                utils::formatUL variable        utils
#> 82                            utils::getAnywhere variable        utils
#> 132                        utils::getCRANmirrors variable        utils
#> 111                      utils::getFromNamespace variable        utils
#> 152                          utils::getParseData variable        utils
#> 68                           utils::getParseText variable        utils
#> 144                           utils::getS3method variable        utils
#> 208                       utils::getSrcDirectory variable        utils
#> 163                        utils::getSrcFilename variable        utils
#> 124                        utils::getSrcLocation variable        utils
#> 92                              utils::getSrcref variable        utils
#> 78                      utils::getTxtProgressBar variable        utils
#> 149                               utils::glob2rx variable        utils
#> 85                        utils::globalVariables variable        utils
#> 23                                utils::hasName variable        utils
#> 177                                  utils::head variable        utils
#> 77                            utils::head.matrix variable        utils
#> 164                                  utils::help variable        utils
#> 127                          utils::help.request variable        utils
#> 87                            utils::help.search variable        utils
#> 45                             utils::help.start variable        utils
#> 198                               utils::history variable        utils
#> 202                            utils::hsearch_db variable        utils
#> 123                   utils::hsearch_db_concepts variable        utils
#> 98                    utils::hsearch_db_keywords variable        utils
#> 41                       utils::install.packages variable        utils
#> 166                    utils::installed.packages variable        utils
#> 25                          utils::is.relistable variable        utils
#> 118                            utils::isS3method variable        utils
#> 69                         utils::isS3stdGeneric variable        utils
#> 63                          utils::limitedLabels variable        utils
#> 147                           utils::loadhistory variable        utils
#> 135                       utils::localeToCharset variable        utils
#> 134                                utils::ls.str variable        utils
#> 125                               utils::lsf.str variable        utils
#> 71                             utils::maintainer variable        utils
#> 180                    utils::make.packages.html variable        utils
#> 207                           utils::make.socket variable        utils
#> 54              utils::makeRweaveLatexCodeRunner variable        utils
#> 186                          utils::memory.limit variable        utils
#> 194                           utils::memory.size variable        utils
#> 195                                  utils::menu variable        utils
#> 218                               utils::methods variable        utils
#> 169                           utils::mirror2html variable        utils
#> 70                             utils::modifyList variable        utils
#> 16                           utils::new.packages variable        utils
#> 42                                   utils::news variable        utils
#> 59                                    utils::nsl variable        utils
#> 84                            utils::object.size variable        utils
#> 150                          utils::old.packages variable        utils
#> 46                              utils::osVersion variable        utils
#> 88                       utils::package.skeleton variable        utils
#> 120                           utils::packageDate variable        utils
#> 104                    utils::packageDescription variable        utils
#> 155                           utils::packageName variable        utils
#> 113                         utils::packageStatus variable        utils
#> 86                         utils::packageVersion variable        utils
#> 182                                  utils::page variable        utils
#> 39                                 utils::person variable        utils
#> 18                             utils::personList variable        utils
#> 188                                  utils::pico variable        utils
#> 66                         utils::process.events variable        utils
#> 9                                  utils::prompt variable        utils
#> 161                            utils::promptData variable        utils
#> 178                          utils::promptImport variable        utils
#> 40                          utils::promptPackage variable        utils
#> 49                           utils::rc.getOption variable        utils
#> 74                             utils::rc.options variable        utils
#> 96                            utils::rc.settings variable        utils
#> 5                               utils::rc.status variable        utils
#> 196                              utils::read.csv variable        utils
#> 114                             utils::read.csv2 variable        utils
#> 13                             utils::read.delim variable        utils
#> 20                            utils::read.delim2 variable        utils
#> 122                              utils::read.DIF variable        utils
#> 173                          utils::read.fortran variable        utils
#> 60                               utils::read.fwf variable        utils
#> 17                            utils::read.socket variable        utils
#> 3                              utils::read.table variable        utils
#> 64                       utils::readCitationFile variable        utils
#> 112                               utils::recover variable        utils
#> 133                                utils::relist variable        utils
#> 151                       utils::remove.packages variable        utils
#> 179                          utils::removeSource variable        utils
#> 34                                  utils::Rprof variable        utils
#> 107                              utils::Rprofmem variable        utils
#> 11                               utils::RShowDoc variable        utils
#> 121                           utils::RSiteSearch variable        utils
#> 99                                  utils::rtags variable        utils
#> 75                                utils::Rtangle variable        utils
#> 157                          utils::RtangleSetup variable        utils
#> 159                       utils::RtangleWritedoc variable        utils
#> 181                     utils::RweaveChunkPrefix variable        utils
#> 62                      utils::RweaveEvalWithOpt variable        utils
#> 103                           utils::RweaveLatex variable        utils
#> 81                      utils::RweaveLatexFinish variable        utils
#> 7                      utils::RweaveLatexOptions variable        utils
#> 29                       utils::RweaveLatexSetup variable        utils
#> 185                   utils::RweaveLatexWritedoc variable        utils
#> 141                         utils::RweaveTryStop variable        utils
#> 139                           utils::savehistory variable        utils
#> 130                           utils::select.list variable        utils
#> 35                            utils::sessionInfo variable        utils
#> 76                          utils::setBreakpoint variable        utils
#> 31                        utils::setRepositories variable        utils
#> 90                      utils::setTxtProgressBar variable        utils
#> 93                                  utils::stack variable        utils
#> 211                               utils::Stangle variable        utils
#> 172                                   utils::str variable        utils
#> 217                            utils::strcapture variable        utils
#> 102                            utils::strOptions variable        utils
#> 89                           utils::summaryRprof variable        utils
#> 209                  utils::suppressForeignCheck variable        utils
#> 58                                 utils::Sweave variable        utils
#> 22                            utils::SweaveHooks variable        utils
#> 154                     utils::SweaveSyntaxLatex variable        utils
#> 105                     utils::SweaveSyntaxNoweb variable        utils
#> 15                         utils::SweaveSyntConv variable        utils
#> 79                                   utils::tail variable        utils
#> 38                            utils::tail.matrix variable        utils
#> 126                                   utils::tar variable        utils
#> 91                              utils::timestamp variable        utils
#> 27                               utils::toBibtex variable        utils
#> 165                               utils::toLatex variable        utils
#> 160                        utils::txtProgressBar variable        utils
#> 44                           utils::type.convert variable        utils
#> 210                           utils::undebugcall variable        utils
#> 168                               utils::unstack variable        utils
#> 204                                 utils::untar variable        utils
#> 57                                  utils::unzip variable        utils
#> 148                       utils::update.packages variable        utils
#> 10                                utils::upgrade variable        utils
#> 197                              utils::url.show variable        utils
#> 4                               utils::URLdecode variable        utils
#> 43                              utils::URLencode variable        utils
#> 2                                      utils::vi variable        utils
#> 200                                  utils::View variable        utils
#> 143                              utils::vignette variable        utils
#> 32                            utils::warnErrList variable        utils
#> 6                               utils::write.csv variable        utils
#> 193                            utils::write.csv2 variable        utils
#> 56                           utils::write.socket variable        utils
#> 19                            utils::write.table variable        utils
#> 153                                 utils::xedit variable        utils
#> 216                                utils::xemacs variable        utils
#> 206                                   utils::zip variable        utils

# A partial identifier
completion("item3 <- va", description = TRUE)

#>       completion     type                                            desc
#> 1       validEnc function  Check if a Character Vector is Validly Encoded
#> 2    validObject function                  Test the Validity of an Object
#> 3 validSlotNames function        Utilities for Managing Class Definitions
#> 4      validUTF8 function  Check if a Character Vector is Validly Encoded
#> 5         vapply function          Apply a Function over a List or Vector
#> 6            var function Correlation, Variance and Covariance (Matrices)
#> 7       var.test function                 F Test to Compare Two Variances
#> 8 variable.names function        Case and Variable Names of Fitted Models
#> 9        varimax function            Rotation Methods for Factor Analysis
#>   context
#> 1    base
#> 2 methods
#> 3 methods
#> 4    base
#> 5    base
#> 6   stats
#> 7   stats
#> 8   stats
#> 9   stats

# Otherwise, a list with the content of .GlobalEnv
completion("item4 <- ")

#>   completion     type
#> 1       iris variable

# TODO: directory and filename completion!
rm(iris, t1)

#> Warning: object 'iris' not found