Execute a command in the GUI client

This function is not intended to be used at the command line (except for debugging purposes). It executes a command string to a (compatible) GUI client.

gui_cmd(command, ...)

gui_load(...)

gui_source(...)

gui_save(...)

gui_import(...)

gui_export(...)

gui_report(...)

gui_setwd(...)

guiCmd(command, ...)

guiLoad(...)

guiSource(...)

guiSave(...)

guiImport(...)

guiExport(...)

guiReport(...)

guiSetwd(...)

Arguments

command
The command string to execute in the GUI client.

...
Parameters provided for the command to execute in the GUI client.

Value

The result of the command if it succeed, or NULL if the command cannot be run (i.e., .guiCmd() is not defined in SciViews:TempEnv).

Details

You must define a function .guiCmd() in the SciViews:TempEnv environment that will take first argument as the name of the command to execute (like source, save, import, etc.), and ... with arguments to the command to send. Depending on your GUI, you should have code that delegates the GUI elements (ex: display a dialog asking for a .Rdata file to source) and then, execute the command in R with the selected file as attribute.

See also