List all methods associated with a generic function or a class, or all types associated with a method

List all S3 and/or S4 methods for a generic function or for a class. List all types for a method; types are variants for a given method defined in a way it is easy to add other variants dynamically (on the contrary to a usual type = or which = argument, like in plot.ts() or plot.lm(), respectively.

list_methods(
  f = character(),
  class = NULL,
  S3 = TRUE,
  S4 = TRUE,
  mixed = TRUE,
  filter = getOption("svGUI.methods")
)

list_types(method, class = "default", strict = FALSE)

listMethods(
  f = character(),
  class = NULL,
  S3 = TRUE,
  S4 = TRUE,
  mixed = TRUE,
  filter = getOption("svGUI.methods")
)

listTypes(method, class = "default", strict = FALSE)

Arguments

f
The name of the generic function (character string), used only when class = NULL.

class
The name of a class.

S3
If TRUE, list of S3 methods.

S4
If TRUE, list of S4 methods.

mixed
If TRUE, S3 and S4 methods are mixed together in a character vector, otherwise, S3 and S4 methods are reported separately in a list.

filter
A list of methods to consider when listing class methods. Only classes in this list that are defined for the class are returned. Store the list of methods you want in the options "svGUI.methods". The package proposes a reasonable starting point on loading if this option is not defined yet.

method
The method name.

strict
Do we list only types for the class (TRUE), or all possible types, including for inherited objects, and default ones FALSE, by default)?

Value

For list_methods(), if mixed = TRUE, a list with components:

S3 The S3 methods for the generic function or the class, or character(0) if none

S4 The S4 methods for the generic function or the class, or character(0) if none.

Otherwise, a character vector with the requested methods.

For list_types(), a vector with character strings with methods' type names.

Note

list_types() is only useful for special generic functions with type argument like view, copy or export. These functions offer a mechanism to easily add custom types, and the present function list them. For S3 objects a type is simply a function whose name is : method_type.class. So, adding new types for method is very easy to implement.

See also

Examples

# Generic functions
list_methods("t.test")               # S3

#> [1] "x=\"default\"" "x=\"formula\""

list_methods("show", mixed = FALSE)     # S4

#> $S3
#> character(0)
#> 
#> $S4
#>  [1] "object=\"ANY\""                      
#>  [2] "object=\"C++Class\""                 
#>  [3] "object=\"C++Function\""              
#>  [4] "object=\"C++Object\""                
#>  [5] "object=\"classGeneratorFunction\""   
#>  [6] "object=\"classRepresentation\""      
#>  [7] "object=\"envRefClass\""              
#>  [8] "object=\"externalRefMethod\""        
#>  [9] "object=\"genericFunction\""          
#> [10] "object=\"genericFunctionWithTrace\"" 
#> [11] "object=\"MethodDefinition\""         
#> [12] "object=\"MethodDefinitionWithTrace\""
#> [13] "object=\"MethodSelectionReport\""    
#> [14] "object=\"MethodWithNext\""           
#> [15] "object=\"MethodWithNextWithTrace\""  
#> [16] "object=\"Module\""                   
#> [17] "object=\"namedList\""                
#> [18] "object=\"ObjectsWithPackage\""       
#> [19] "object=\"oldClass\""                 
#> [20] "object=\"refClassRepresentation\""   
#> [21] "object=\"refMethodDef\""             
#> [22] "object=\"refObjectGenerator\""       
#> [23] "object=\"signature\""                
#> [24] "object=\"sourceEnvironment\""        
#> [25] "object=\"traceable\""                
#> 

list_methods("ls") # None, not a generic function!

#> character(0)

# Classes
# Only the following methods are considered
getOption("gui.methods")

#> NULL

list_methods(class = "data.frame")

#> [1] "formula" "head"    "plot"    "summary" "tail"   

list_methods(class = "lm")

#> [1] "anova"     "confint"   "formula"   "logLik"    "plot"      "predict"  
#> [7] "residuals" "summary"   "vcov"     

# List method types
list_types("view")  # All default view types currently defined

#> character(0)

list_types("view", "data.frame")

#> character(0)

list_types("view", "data.frame", TRUE) # None, except if you defined custom views!

#> character(0)