SciViews - Miscellaneous functions

Miscellaneous functions for SciViews or general use. The svMisc package collects together a series of functions that are shared with svXXX packages.

Important functions

temp_env() for using a temporary environment attached to the search path,

temp_var() create the name of temporary variables,

capture_all() to capture R output, errors, warnings and messages,

parse_text() to parse any R expression, including partial or incorrect ones (fails gracefully).